+++ src/main/java/com/example/vitamate/service/SupplementService/SupplementCommandService.java
@@ -1,14 +1,14 @@
 package com.example.vitamate.service.SupplementService;
 
+import com.example.vitamate.web.dto.ReviewRequestDTO;
+import com.example.vitamate.web.dto.ReviewResponseDTO;
 import com.example.vitamate.web.dto.SupplementRequestDTO;
 import com.example.vitamate.web.dto.SupplementResponseDTO;
-
-import java.time.LocalDate;
 
 public interface SupplementCommandService {
     SupplementResponseDTO.AddIntakeSupplementResultDTO addIntakeSupplement(String email, Long supplementId, SupplementRequestDTO.AddIntakeSupplementDTO requestDTO);
 
     SupplementResponseDTO.AddScrapResultDTO addScrap(String email, Long supplementId);
     SupplementResponseDTO.DeleteScrapResultDTO deleteScrap(String email, Long supplementId);
-    SupplementResponseDTO.ReviewResultDTO addReview(String email, Long supplementId, SupplementRequestDTO.AddReviewDTO requestDTO);
+    ReviewResponseDTO.ReviewResultDTO addReview(String email, Long supplementId, ReviewRequestDTO.AddReviewDTO requestDTO);
 }
