+++ src/main/java/com/example/vitamate/web/dto/ReviewResponseDTO.java
@@ -1,0 +1,22 @@
+package com.example.vitamate.web.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+public class ReviewResponseDTO {
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class ReviewResultDTO{
+        Long reviewId;
+        String nickname;
+        String content;
+        Integer grade;
+        LocalDateTime createdDate;
+    }
+}
